requirements.txt

streamlit
rdkit-pypi
pandas
openpyxl
xlsxwriter
